+++ js/main.js
@@ -3,3 +3,4 @@ import './preview.js';
 import './user-form.js';
 import './validation.js';
 import './photo-effects.js';
+import './photo-scale.js';
